pumpMultiplier and potassium accumulation save
Instead of turning the pump entirely on or off for the central compartments a pumpMultiplier can be set to change the max conductance of the central compartments (setting this to 0 turns them off). This allows for testing partial pump block.
Additions to the save_to_file code allow for saving the internal and external potassium concentration of all compartments into a two files.