Fix pump adjustment
The offset calculations to determine which pump should be turned off were incorrect. Previously certain compartments on the lower end of the range that *should* have been set were not being set, and compartments on the higher end of the range that should *not* have been set were being set.
The code now adjusts the pump conduction value for all compartments between node 45 and 55.