added custom electrode code
Testing has validated that the DC electrode functions the exact same with the previous configuration and the custom electrode configuration.
Validation was done with a -670000nA stimulation 1mm from the center of the axon and the voltage values were saved to file and compared to show custom electrodes function the exact same.